scripts/requirements.txt
opencv-python==4.10.0.84
numpy==1.24.0
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
transformers==4.45.2
accelerate==1.1.1